Module aws_sdk_athena::error

Modules

internal_server_exception

See InternalServerException

invalid_request_exception

See InvalidRequestException

metadata_exception

See MetadataException

resource_not_found_exception

See ResourceNotFoundException

too_many_requests_exception

See TooManyRequestsException

Structs

BatchGetNamedQueryError

BatchGetQueryExecutionError

CreateDataCatalogError

CreateNamedQueryError

CreatePreparedStatementError

CreateWorkGroupError

DeleteDataCatalogError

DeleteNamedQueryError

DeletePreparedStatementError

DeleteWorkGroupError

GetDataCatalogError

GetDatabaseError

GetNamedQueryError

GetPreparedStatementError

GetQueryExecutionError

GetQueryResultsError

GetTableMetadataError

GetWorkGroupError

InternalServerException

Indicates a platform issue, which may be due to a transient condition or

InvalidRequestException

Indicates that something is wrong with the input to the request. For example, a

ListDataCatalogsError

ListDatabasesError

ListEngineVersionsError

ListNamedQueriesError

ListPreparedStatementsError

ListQueryExecutionsError

ListTableMetadataError

ListTagsForResourceError

ListWorkGroupsError

MetadataException

An exception that Athena received when it called a custom metastore.

ResourceNotFoundException

A resource, such as a workgroup, was not found.

StartQueryExecutionError

StopQueryExecutionError

TagResourceError

TooManyRequestsException

Indicates that the request was throttled.

UntagResourceError

UpdateDataCatalogError

UpdatePreparedStatementError

UpdateWorkGroupError

Enums

BatchGetNamedQueryErrorKind

BatchGetQueryExecutionErrorKind

CreateDataCatalogErrorKind

CreateNamedQueryErrorKind

CreatePreparedStatementErrorKind

CreateWorkGroupErrorKind

DeleteDataCatalogErrorKind

DeleteNamedQueryErrorKind

DeletePreparedStatementErrorKind

DeleteWorkGroupErrorKind

GetDataCatalogErrorKind

GetDatabaseErrorKind

GetNamedQueryErrorKind

GetPreparedStatementErrorKind

GetQueryExecutionErrorKind

GetQueryResultsErrorKind

GetTableMetadataErrorKind

GetWorkGroupErrorKind

ListDataCatalogsErrorKind

ListDatabasesErrorKind

ListEngineVersionsErrorKind

ListNamedQueriesErrorKind

ListPreparedStatementsErrorKind

ListQueryExecutionsErrorKind

ListTableMetadataErrorKind

ListTagsForResourceErrorKind

ListWorkGroupsErrorKind

StartQueryExecutionErrorKind

StopQueryExecutionErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateDataCatalogErrorKind

UpdatePreparedStatementErrorKind

UpdateWorkGroupErrorKind