Module aws_sdk_athena::model

Modules

column

See Column

column_info

See ColumnInfo

data_catalog

See DataCatalog

data_catalog_summary

See DataCatalogSummary

database

See Database

datum

See Datum

encryption_configuration

See EncryptionConfiguration

engine_version

See EngineVersion

named_query

See NamedQuery

prepared_statement

See PreparedStatement

prepared_statement_summary

See PreparedStatementSummary

query_execution

See QueryExecution

query_execution_context

See QueryExecutionContext

query_execution_statistics

See QueryExecutionStatistics

query_execution_status

See QueryExecutionStatus

result_configuration

See ResultConfiguration

result_configuration_updates

See ResultConfigurationUpdates

result_set

See ResultSet

result_set_metadata

See ResultSetMetadata

row

See Row

table_metadata

See TableMetadata

tag

See Tag

unprocessed_named_query_id

See UnprocessedNamedQueryId

unprocessed_query_execution_id

See UnprocessedQueryExecutionId

work_group

See WorkGroup

work_group_configuration

See WorkGroupConfiguration

work_group_configuration_updates

See WorkGroupConfigurationUpdates

work_group_summary

See WorkGroupSummary

Structs

Column

Contains metadata for a column in a table.

ColumnInfo

Information about the columns in a query execution result.

DataCatalog

Contains information about a data catalog in an Amazon Web Services account.

DataCatalogSummary

The summary information for the data catalog, which includes its name and type.

Database

Contains metadata information for a database in a data catalog.

Datum

A piece of data (a field in the table).

EncryptionConfiguration

If query results are encrypted in Amazon S3, indicates the encryption option

EngineVersion

The Athena engine version for running queries.

NamedQuery

A query, where QueryString is the list of SQL query statements that

PreparedStatement

A prepared SQL statement for use with Athena.

PreparedStatementSummary

The name and last modified time of the prepared statement.

QueryExecution

Information about a single instance of a query execution.

QueryExecutionContext

The database and data catalog context in which the query execution occurs.

QueryExecutionStatistics

The amount of data scanned during the query execution and the amount of time that it

QueryExecutionStatus

The completion date, current state, submission time, and state change reason (if

ResultConfiguration

The location in Amazon S3 where query results are stored and the encryption

ResultConfigurationUpdates

The information about the updates in the query results, such as output location and

ResultSet

The metadata and rows that comprise a query result set. The metadata describes the

ResultSetMetadata

The metadata that describes the column structure and data types of a table of query

Row

The rows that comprise a query result table.

TableMetadata

Contains metadata for a table.

Tag

A label that you assign to a resource. In Athena, a resource can be a

UnprocessedNamedQueryId

Information about a named query ID that could not be processed.

UnprocessedQueryExecutionId

Describes a query execution that failed to process.

WorkGroup

A workgroup, which contains a name, description, creation time, state, and other

WorkGroupConfiguration

The configuration of the workgroup, which includes the location in Amazon S3

WorkGroupConfigurationUpdates

The configuration information that will be updated for this workgroup, which includes

WorkGroupSummary

The summary information for the workgroup, which includes its name, state,

Enums

ColumnNullable

NOTE: ColumnNullable::Unknown has been renamed to ::UnknownValue.

DataCatalogType

EncryptionOption

QueryExecutionState

StatementType

ThrottleReason

The reason for the query throttling, for example, when it exceeds the concurrent query

WorkGroupState