Module aws_sdk_athena::operation

Structs

BatchGetNamedQuery

Returns the details of a single named query or a list of up to 50 queries, which you

BatchGetQueryExecution

Returns the details of a single query execution or a list of up to 50 query

CreateDataCatalog

Creates (registers) a data catalog with the specified name and properties. Catalogs

CreateNamedQuery

Creates a named query in the specified workgroup. Requires that you have access to the

CreatePreparedStatement

Creates a prepared statement for use with SQL queries in Athena.

CreateWorkGroup

Creates a workgroup with the specified name.

DeleteDataCatalog

Deletes a data catalog.

DeleteNamedQuery

Deletes the named query if you have access to the workgroup in which the query was

DeletePreparedStatement

Deletes the prepared statement with the specified name from the specified

DeleteWorkGroup

Deletes the workgroup with the specified name. The primary workgroup cannot be

GetDataCatalog

Returns the specified data catalog.

GetDatabase

Returns a database object for the specified database and data catalog.

GetNamedQuery

Returns information about a single query. Requires that you have access to the

GetPreparedStatement

Retrieves the prepared statement with the specified name from the specified

GetQueryExecution

Returns information about a single execution of a query if you have access to the

GetQueryResults

Streams the results of a single query execution specified by

GetTableMetadata

Returns table metadata for the specified catalog, database, and table.

GetWorkGroup

Returns information about the workgroup with the specified name.

ListDataCatalogs

Lists the data catalogs in the current Amazon Web Services account.

ListDatabases

Lists the databases in the specified data catalog.

ListEngineVersions

Returns a list of engine versions that are available to choose from, including the

ListNamedQueries

Provides a list of available query IDs only for queries saved in the specified

ListPreparedStatements

Lists the prepared statements in the specfied workgroup.

ListQueryExecutions

Provides a list of available query execution IDs for the queries in the specified

ListTableMetadata

Lists the metadata for the tables in the specified data catalog database.

ListTagsForResource

Lists the tags associated with an Athena workgroup or data catalog

ListWorkGroups

Lists available workgroups for the account.

StartQueryExecution

Runs the SQL query statements contained in the Query. Requires you to

StopQueryExecution

Stops a query execution. Requires you to have access to the workgroup in which the

TagResource

Adds one or more tags to an Athena resource. A tag is a label that you

UntagResource

Removes one or more tags from a data catalog or workgroup resource.

UpdateDataCatalog

Updates the data catalog that has the specified name.

UpdatePreparedStatement

Updates a prepared statement.

UpdateWorkGroup

Updates the workgroup with the specified name. The workgroup's name cannot be